+++ gears/src/test/java/org/hortonmachine/gears/modules/TestRasterResolutionResampler.java
@@ -35,7 +35,7 @@ public class TestRasterResolutionResampler extends HMTestCase {
 
     public void testRasterResolutionResampler() throws Exception {
 
-        double[][] flowData = HMTestMaps.flowData;
+        int[][] flowData = HMTestMaps.flowData;
         HashMap<String, Double> envelopeParams = HMTestMaps.getEnvelopeparams();
         CoordinateReferenceSystem crs = HMTestMaps.getCrs();
         GridCoverage2D flowCoverage = CoverageUtilities.buildCoverage("flow", flowData, envelopeParams, crs, true);
